requirements.txt
beautifulsoup4 @ file:///C:/b/abs_0agyz1wsr4/croot/beautifulsoup4-split_1681493048687/work
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
filelock==3.9.0
fonttools==4.46.0
fsspec==2023.4.0
gensim==4.3.2
graphviz==0.20.1
idna==3.4
Jinja2==3.1.2
joblib @ file:///C:/b/abs_1anqjntpan/croot/joblib_1685113317150/work
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.2
mkl-fft @ file:///C:/b/abs_19i1y8ykas/croot/mkl_fft_1695058226480/work
mkl-random @ file:///C:/b/abs_edwkj1_o69/croot/mkl_random_1695059866750/work
mkl-service==2.4.0
mpmath==1.3.0
networkx==3.0
nltk==3.8.1
numpy==1.24.1
packaging==23.2
pandas==2.1.3
Pillow==9.3.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
regex==2023.12.25
requests==2.28.1
scikit-learn==1.3.2
scipy==1.11.4
six==1.16.0
smart-open==6.4.0
soupsieve @ file:///C:/b/abs_bbsvy9t4pl/croot/soupsieve_1696347611357/work
sympy==1.12
threadpoolctl @ file:///Users/ktietz/demo/mc3/conda-bld/threadpoolctl_1629802263681/work
torch==2.1.1+cu121
torchaudio==2.1.1+cu121
torchdata==0.7.1
torchtext==0.16.1
torchvision==0.16.1+cu121
torchviz==0.0.2
tqdm==4.66.1
typing_extensions==4.4.0
tzdata==2023.3
urllib3==1.26.13
